+++ hireHero-app/src/server/controllers/userController.ts
@@ -1,18 +1,13 @@
 import { RequestHandler } from 'express';
 
+import db from '../model/interviewModel';
+
+
 type UserController = {
   queryTimes: RequestHandler;
-  numOfRows: RequestHandler;
-  topCalls: RequestHandler;
-  dbStats: RequestHandler;
-  cacheHitRatio: RequestHandler;
+
 };
-type queryData = {
-  all: any[];
-  mean: number;
-  median: number;
-  slowestQueries: any[];
-} | null;
+
 
 const userController: UserController = {
