+++ hireHero-app/src/server/index.ts
@@ -1,53 +1,75 @@
-import path from "path";
-import type { Request, Response, NextFunction, Router, Express } from "express";
-import express from "express";
+import path from 'path';
+import userApi from './routes/userApi'; /* userApi */
+import interviewApi from './routes/interviewApi'; /* interviewApi */
+import authApi from './routes/authApi'; /* authApi */
+
+/* import Express and its types */
+
+import express, {
+  json,
+  urlencoded,
+  Request,
+  Response,
+  NextFunction,
+} from 'express';
+
+/* initialize express */
+
 const app = express();
-//import apiRouter from './routes/api';
 
-const PORT = 3000;
+/* setup the PORT */
 
-/**
- * handle parsing request body
- */
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
+const port = 3000;
+
+
+/* handles the parsing of the request bodies */
+
+app.use(json());
+app.use(urlencoded({ extended: true }));
+
 
 /**
  * handle requests for static files
  */
 app.use(express.static(path.resolve(__dirname, "../../dist")));
 
-/**
- * define route handlers
- */
-// app.use('/api', apiRouter);
+//define route handlers
 
-// catch-all route handler for any requests to an unknown route
-app.use((req, res) =>
-  res.status(404).send("This is not the page you're looking for...")
-);
 
-/**
- * express error handler
- * @see https://expressjs.com/en/guide/error-handling.html#writing-error-handlers
- */
+// route for all Postgres Interview Requests
 
-app.use((err: any, req: Request, res: Response, next: NextFunction) => {
-  const defaultErr = {
-    log: "Express error handler caught unknown middleware error",
+app.use('/api/user', userApi);
+app.use('/api/interviews', interviewApi);
+app.use('/api/auth', authApi);
+
+
+// Catch-all route handler
+
+app.use('*', (req: Request, res: Response) => {
+
+  res.sendStatus(404);
+
+});
+
+/* Global Error Handler */
+
+app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
+  const defaultError = {
+    log: 'Express error handler caught unknown middleware error',
     status: 500,
-    message: { err: "An error occurred" },
+    message: { Error: 'An error occurred' },
   };
-  const errorObj = Object.assign({}, defaultErr, err);
-  console.log(errorObj.log);
+
+  const errorObj = Object.assign({}, defaultError, err);
   return res.status(errorObj.status).json(errorObj.message);
 });
 
-/**
- * start server
- */
-app.listen(PORT, () => {
-  console.log(`Server listening on port: ${PORT}...`);
+/* start the server */
+
+const server = app.listen(port, () => {
+  console.log(`Server started on Port: ${port}`);
 });
 
-module.exports = app;
+/* export the server */
+
+module.exports = server;
