+++ hireHero-app/src/server/model/interviewModel.ts
@@ -1,7 +1,19 @@
 const { Pool } = require('pg');
 
+/* import Express and its types */
+
+
+import express, {
+  json,
+  urlencoded,
+  Request,
+  Response,
+  NextFunction,
+} from 'express';
+
+
 // URI for Elephant SQL (If not using this one then you will need to manually setup some tables for everything to work)
-const PG_URI = '< enter URI >';
+const PG_URI = 'postgres://qoewcqtp:zbJbMieoG08RobHp3xgRukVNMQ7NuQ9v@chunee.db.elephantsql.com/qoewcqtp';
 
 const pool = new Pool({
   connectionString: PG_URI
@@ -9,11 +21,12 @@ const pool = new Pool({
 
 module.exports = {
 
-  query: (text, params, callback) => {
+  query: (text: string, params?: Array<string>, callback?: any) => {
 
     console.log('executed query', text);
 
     return pool.query(text, params, callback);
     
   }
+  
 };
